refactor/#71: 채팅 하단 패널에서 갤러리 및 카메라 버튼 핸들러 통합
